// dashboard/src/utils/hasura-api/generated/schemas/cronTrigger.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { CronTriggerPayload } from './cronTriggerPayload';
import type { Headers } from './headers';
import type { RequestTransformation } from './requestTransformation';
import type { ResponseTransformation } from './responseTransformation';
import type { RetryConfCT } from './retryConfCT';

export interface CronTrigger {
  /** Name of the cron trigger */
  name: string;
  /** URL of the webhook. A String value which supports templating environment variables enclosed in {{ and }}. */
  webhook: string;
  /** Cron expression at which the trigger should be invoked. */
  schedule: string;
  /** Any JSON payload which will be sent when the webhook is invoked. */
  payload?: CronTriggerPayload;
  /** List of headers to be sent with the webhook */
  headers?: Headers;
  /** Retry configuration if scheduled invocation delivery fails */
  retry_conf?: RetryConfCT;
  /** Flag to indicate whether a trigger should be included in the metadata. When a cron trigger is included in the metadata, the user will be able to export it when the Metadata of the graphql-engine is exported. */
  include_in_metadata: boolean;
  /** Custom comment. */
  comment?: string;
  request_transform?: RequestTransformation;
  response_transform?: ResponseTransformation;
}